Don't set activity completed on not inscribed courses
This function was intended to better indicate course progress:
If a user was not inscribed at all to a subcourse and if this activity should be hidden when a user was not inscribed, we have set the activity to completed (as it was not visible anyways). But this lead to problems on the booking elective, where users only later subscribed to courses, but the activity was completed before.